api/genToken.py
import jwt
from jwt import ExpiredSignatureError, InvalidTokenError
from fastapi import HTTPException

# Generar un token JWT
def createToken(datos: dict):
    token: str = jwt.encode(payload=datos, key='secretkey', algorithm='HS256')
    return token

# Validar token JWT
def validateToken(token: str):
    try:
        data: dict = jwt.decode(token, key='secretkey', algorithms=['HS256'])
        return data
    except ExpiredSignatureError:
        raise HTTPException(status_code=403, detail='Token expirado')
    except InvalidTokenError:
        raise HTTPException(status_code=403, detail='Token no autorizado')