api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
email-validator>=1.3.1
pydantic[email]>=2.0
PyJWT>=2.0.0 
mysql-connector-python>=9.2.0